rn-smo/install/demo/ui-visualization/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
kafka-python==2.0.2
websockets==12.0
httpx==0.27.0

